+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - Conversion tool: added parser for Hotbit exchange.
 - Conversion tool: added parser for Liquid exchange.
 - Conversion tool: added parser for Energy Web explorer.
+- Qt Wallet parser: recognise Namecoin operations.
 ### Changed
 - Sort wallet names in audit debug as case-insensitive.
 - Data source names in config are now case-insensitive.